requirements.txt
Flask
pillow
watchdog
setuptools
gunicorn
python-ffmpeg
